Fix incorrect accel values and data type

+++ firmware/src/globals.h
@@ -8,7 +8,7 @@
 
 
 extern uint8_t virtual_leds[NUM_PIXELS][3];
-extern uint8_t accel_data[3];
+extern float accel_data[3];
 
 extern LEDFilter_Smooth led_filter_smooth;
 extern LEDFilter_Basic led_filter_basic;
